HACK: Panda5: Increase dalvik heap size
We have seen ANR's caused by out of memory errors in the dalvik heap.
I've been researching the right heap size to use for this board but
so far have not found a lot of information out there on how to
calculate the proper size given different board parameters (DDR size,
screen resolution, etc...).

So, for now I am going to increase it to the max pre-configured heap
size available on Android and we'll continue to investigate whether
an additional increase will be needed.

Change-Id: Ice7ca864576321e816f467ca371b19a4a591c9c5
Signed-off-by: Kevin McCray <kevin.mccray@ti.com>